bugs fix endpoint get brands

+++ dsf/payment/service/request/brand_request.go
@@ -2,8 +2,8 @@ package request
 
 type BrandsRequest struct {
 	Keyword string `form:"keyword"`
-	Limit   int64  `form:"limit"`
-	Offset  int64  `form:"offset"`
+	Limit   int    `form:"limit"`
+	Offset  int    `form:"offset"`
 }
 
 func (f *BrandsRequest) Validate() error {
